+++ integrations/kubecost/CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- towncrier release notes start -->
 
+# Port_Ocean 0.1.1 (2023-10-26)
+
+### Improvements
+
+- Added a sanity check for the kubecost startup to check for connectivity issues (PORT-5039)
+
+### Bug Fixes
+
+- Fixed the config.yml to read the kubecostHost from the environment variables (PORT-5039)
+
+
 # Kubecost 0.1.0 (2023-10-24)
 
 ### Features